Band of Endurance; Bug Fix
FIX
-Dread of the Super Natural was activating every time the user was affected by any DOT
-The player was leveling up every turn
-Unity would throw an error when a user could not attack

NEW
-Band of Endurance

REMOVED
-Level Stat System (was being unused since added stat growth per level)

+++ Jogo/Game Project/Assets/Code/Utils.cs
@@ -109,17 +109,17 @@ public struct DMG
     }
 
     //used on AI
-    public void Multiply(int num)
+    public void Multiply(float num)
     {
         phyDmg *= num;
         magicDmg *= num;
         trueDmg *= num;
-        sanityDmg *= num;
+        sanityDmg *= (int)num;
 
         heal *= num;
         healMana *= num;
         healStamina *= num;
-        healSanity *= num;
+        healSanity *= (int)num;
 
         shield *= num;
     }
